LayoutTests/platform/chromium/accessibility/multiselect-list-reports-active-option-expected.txt


This tests that navigating in a multiselect list updates selection and the active selected option and sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibleOne.isSelected is true
PASS accessibleOne.isSelectedOptionActive is true
PASS accessibleTwo.isSelected is false
PASS accessibleTwo.isSelectedOptionActive is false
PASS accessibleThree.isSelected is false
PASS accessibleThree.isSelectedOptionActive is false
PASS accessibleOne.isSelected is false
PASS accessibleOne.isSelectedOptionActive is false
PASS accessibleTwo.isSelected is true
PASS accessibleTwo.isSelectedOptionActive is true
PASS accessibleThree.isSelected is false
PASS accessibleThree.isSelectedOptionActive is false
PASS accessibleOne.isSelected is false
PASS accessibleOne.isSelectedOptionActive is false
PASS accessibleTwo.isSelected is true
PASS accessibleTwo.isSelectedOptionActive is false
PASS accessibleThree.isSelected is true
PASS accessibleThree.isSelectedOptionActive is true
List notification: SelectedChildrenChanged
List notification: SelectedChildrenChanged

TEST COMPLETE
